Remove Factory dependency from GOVKit and introduce singletons for CoreData repo and ActivityService
