+++ Production/govuk_ios/Extensions/Container/Container+Repositories.swift
@@ -1,17 +1,26 @@
 import Foundation
 import Factory
+import GOVKit
 import RecentActivity
 
 extension Container {
+    public var activityRepository: Factory<ActivityRepositoryInterface> {
+        Factory(self) {
+            ActivityRepository(
+                coreData: CoreDataRepository.govUK
+            )
+        }
+    }
+
     var topicsRepository: Factory<TopicsRepositoryInterface> {
         Factory(self) {
-            TopicsRepository(coreData: self.coreDataRepository())
+            TopicsRepository(coreData: CoreDataRepository.govUK)
         }
     }
 
     var searchHistoryRepository: Factory<SearchHistoryRepositoryInterface> {
         Factory(self) {
-            SearchHistoryRepository(coreData: self.coreDataRepository())
+            SearchHistoryRepository(coreData: CoreDataRepository.govUK)
         }
     }
 }
