--- RecentActivity/Sources/RecentActivity/Extensions/Container+RecentActivity.swift
@@ -1,13 +1,0 @@
-import Foundation
-import Factory
-import GOVKit
-
-extension Container {
-    public var activityRepository: Factory<ActivityRepositoryInterface> {
-        Factory(self) {
-            ActivityRepository(
-                coreData: self.coreDataRepository()
-            )
-        }
-    }
-}
