+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/Builders/ViewControllerBuilderTests.swift
@@ -72,10 +72,6 @@ struct ViewControllerBuilderTests {
 
     @Test
     func recentActivity_returnsExpectedResult() {
-        let coreData = CoreDataRepository.arrangeAndLoad
-        Container.shared.coreDataRepository.register {
-            coreData
-        }
         let subject = ViewControllerBuilder()
         let result = subject.recentActivity(
             analyticsService: MockAnalyticsService(),
